% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_string.R
\name{api_url}
\alias{api_url}
\alias{api_url.default}
\alias{api_url.query_string}
\title{Extract `api_url` from an object}
\usage{
api_url(x, ...)

\method{api_url}{default}(x, ...)

\method{api_url}{query_string}(x, ...)
}
\arguments{
\item{x}{object from which to extract `api_url`}

\item{...}{ignored}
}
\value{
the `api_url` (if available)
}
\description{
Extract `api_url` from an object
}
