% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{dataVersion}
\alias{dataVersion}
\title{(Canned Query) dataVersion}
\usage{
dataVersion(api_url = OT_API)
}
\arguments{
\item{api_url}{API URL}
}
\value{
API name and dataVersion
}
\description{
(Canned Query) dataVersion
}
