% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{get_schema}
\alias{get_schema}
\title{Get the schema for a GraphQL API}
\usage{
get_schema(api_url = OT_API)
}
\arguments{
\item{api_url}{API query endpoint}
}
\value{
a character string containing the schema, which can be used in
  [query_builder()] without re-fetching
}
\description{
Get the schema for a GraphQL API
}
