% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{OPENTARGETS_API}
\alias{OPENTARGETS_API}
\title{Default API endpoint}
\format{
An object of class \code{character} of length 1.
}
\usage{
OPENTARGETS_API
}
\description{
Default API endpoint
}
\keyword{datasets}
