% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{OT_GENETICS_API}
\alias{OT_GENETICS_API}
\title{Default OpenTargets Genetics API endpoint}
\format{
An object of class \code{character} of length 1.
}
\usage{
OT_GENETICS_API
}
\description{
Default OpenTargets Genetics API endpoint
}
\keyword{datasets}
