% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{parse_elements}
\alias{parse_elements}
\title{Extract elements from the schema}
\usage{
parse_elements(el, struct = "type", subquery = FALSE)
}
\arguments{
\item{el}{element to process}
}
\value{
a list containing `object_class` (name of schema element) and
  arguments with associated class, description, and non_nullable flag
}
\description{
Extract elements from the schema
}
\examples{
x <- parse_elements(types[21])
jsonview::json_tree_view(x)
}
