% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_string.R
\name{print.query_string}
\alias{print.query_string}
\title{Print a query_string object}
\usage{
\method{print}{query_string}(x, ...)
}
\arguments{
\item{x}{query_string object}

\item{...}{ignored}
}
\value{
the query_string object
}
\description{
Print a query_string object
}
