% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{query_builder}
\alias{query_builder}
\title{Build a GraphQL Query String}
\usage{
query_builder(api_url = NULL)
}
\arguments{
\item{api_url}{URL of the API from which to read schema}
}
\value{
a query string, possibly to be used in [run_query()]
}
\description{
Build a GraphQL Query String
}
