% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{run_query}
\alias{run_query}
\title{Run a GraphQL query}
\usage{
run_query(query_string, variables, api_url = OPENTARGETS_API)
}
\arguments{
\item{query_string}{query string, possibly built with [query_builder()]}

\item{variables}{`list` of variables to be used in the query}
}
\value{
returned data from the API
}
\description{
Run a GraphQL query
}
