% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{query_builder}
\alias{query_builder}
\title{Build a GraphQL Query String}
\usage{
query_builder(schema = NULL, api_url = OT_API)
}
\arguments{
\item{schema}{a local copy of the schema as a character string}

\item{api_url}{URL of the API from which to read schema. By default this will
use the OpenTargets API}
}
\value{
a query string, possibly to be used in [run_query()]
}
\description{
Build a GraphQL Query String
}
